404 error handling, templates are now working as templates

+++ src/main/java/cz/czechitas/java2webapps/ukol6/repository/VizitkaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import cz.czechitas.java2webapps.ukol6.entity.Vizitka;
 
 @Repository
-public interface VizitkaRepository extends CrudRepository<Vizitka, Long> {
+public interface VizitkaRepository extends CrudRepository<Vizitka, Integer> {
 }
